Split crypto.rs into multiple files
