+++ .vscode/settings.json
@@ -4,9 +4,14 @@
     "Cdecl",
     "chrono",
     "cloc",
+    "decryptable",
     "dylib",
+    "encryptable",
     "Hkdf",
+    "Hmac",
     "Maybeable",
+    "Oaep",
+    "Pbkdf",
     "schemars",
     "uniffi",
     "wordlist"
